update to ebooks _backup script
Added desc block.

Added -vps param for backing up to remove server. Includes error handling in event empty list is passed to compare-object.

All routines log to file.

+++ powershell/backup_ebooks/_backup.ps1
@@ -1,17 +1,19 @@
-# Script to backup ebooks to usb drive. Script will prompt for destination drive letter
-# Creates a comparison list to ensure only ebooks not previously backed up are copied.
-# Error handling in case destination is empty will simply copy all available ebooks to
-# flash pen. Script assumes it's in your ebooks directory otherwise set path of source.
-# Script ignores any files in a folder named 'My Kindle Content' (duplicates).
-
-# Feel free to use this script or modify it for your own purposes. I take no responsibility
-# for use of this script and I don't make any promises that it will be suitable for any
-# specific purpose. I find it useful for my own needs.
-
-$usb_drive = Read-Host -Prompt 'Drive letter of usb flashpen?'
+<#
+	This script allows me to backup ebooks to usb and remote server via -vps flag.
+	Both routines check for empty lists passed to comparator object.
+	Will only backup files not previously backed up.
+	Ignores any files in a folder names My Kindle Content as well as extensions not
+	matching the specified exclusions regex
+	Script will prompt for usb drive letter if -vps parameter not passed.
+ 
+	Feel free to use or modify this script for your own purposes. Use at your own risk,
+	I take no responsibility for its use and make no promises that it will be fit for
+	any particular purpose, I find it useful for my own purposes.
+#>
+param([switch]$vps)
 
 $source = "$PWD"
-$destination = "${usb_drive}:\ebooks"
+$exclusions = ".pdf|.epub|.zip"
 
 $logfile = "$PWD\_backup.log"
 
@@ -20,24 +22,58 @@ If (-Not [System.IO.File]::Exists($path)) {
 	Out-File -FilePath $logfile
 }
 
-# full list of files in source and destination
-$list_s = Get-ChildItem -Recurse -path $source
-$list_d = Get-ChildItem -Recurse -path $destination
 
-# difference list
-Try {
-	ForEach ($file in (Compare-Object -ReferenceObject $list_s -DifferenceObject $list_d -PassThru | `
-	Where-Object { $_.PSParentPath -notlike "*My Kindle Content" -and $_.Extension -match ".pdf|.epub|.zip"})) {
+if($vps) {
+	$destination = "~/scp/ebooks2"
+	$_getremotelist = "ssh vps 'ls -A1 ${destination}'"	
+	
+	# full list of files in source and destination
+	$list_s = Get-ChildItem -Recurse -path $source
+	$list_d = Invoke-Expression $_getremotelist
+	
+	Try {
+		ForEach ($file in (Compare-Object -ReferenceObject $list_s -DifferenceObject $list_d -PassThru | `
+		Where-Object { $_.PSParentPath -notlike "*My Kindle Content" -and $_.Extension -match $exclusions})) {
+				"Copying only files that don't exist in destination:" | Tee-Object -Append $logfile
+				"$file to $destination" | Tee-Object -Append $logfile
+				scp -vp $source\$file vps:$destination
+			}
+	} Catch {
+		Write-Host "Either source or destination are empty, failed to generate comparison list"
+	 
+		"Copying ebooks to destination!" | Tee-Object -Append $logfile
+		
+		ForEach ($file in (Get-ChildItem -Recurse -path $source | `
+		Where-Object { $_.PSParentPath -notlike "*My Kindle Content" -and $_.Extension -match $exclusions})) {
 			"Copying only files that don't exist in destination:" | Tee-Object -Append $logfile
 			"$file to $destination" | Tee-Object -Append $logfile
-			Copy-Item $source\$file $destination
-		}
-} Catch {
-	Write-Host "Either source or destination are empty, failed to generate comparison list"
- 
-	"Copying ebooks to destination!" | Tee-Object -Append $logfile
+			scp -vp $source\$file vps:$destination
+		}			
+	}
+			
+} else {
+	$usb_drive = Read-Host -Prompt 'Drive letter of usb flashpen?'
+	$destination = "${usb_drive}:\ebooks"	
 	
-	Get-ChildItem -Recurse -path $source | `
-	Where-Object { $_.PSParentPath -notlike "*My Kindle Content" -and $_.Extension -match ".pdf|.epub|.zip"} | `
-	Copy-Item -Destination $destination -Recurse -Force -verbose | Tee-Object -Append $logfile
+	# full list of files in source and destination
+	$list_s = Get-ChildItem -Recurse -path $source
+	$list_d = Get-ChildItem -Recurse -path $destination
+	
+	# create difference list and copy based on simple regex
+	Try {
+		ForEach ($file in (Compare-Object -ReferenceObject $list_s -DifferenceObject $list_d -PassThru | `
+		Where-Object { $_.PSParentPath -notlike "*My Kindle Content" -and $_.Extension -match $exclusions})) {
+				"Copying only files that don't exist in destination:" | Tee-Object -Append $logfile
+				"$file to $destination" | Tee-Object -Append $logfile
+				Copy-Item $source\$file $destination
+			}
+	} Catch {
+		Write-Host "Either source or destination are empty, failed to generate comparison list"
+	 
+		"Copying ebooks to destination!" | Tee-Object -Append $logfile
+		
+		Get-ChildItem -Recurse -path $source | `
+		Where-Object { $_.PSParentPath -notlike "*My Kindle Content" -and $_.Extension -match $exclusions} | `
+		Copy-Item -Destination $destination -Recurse -Force -verbose | Tee-Object -Append $logfile
+	}		
 }
